+++ CHANGELOG.md
@@ -1,15 +1,32 @@
 # Changelog
 
+## [3.11.1](https://github.com/gitlabform/gitlabform/compare/v3.11.0...v3.11.1) (2024-05-16)
+
+
+### Bug Fixes
+
+* release process and deploy container to GHCR ([741ff75](https://github.com/gitlabform/gitlabform/commit/741ff751fc26f17ccaca51fd8212175e7657839b))
+
+## [3.11.0](https://github.com/gitlabform/gitlabform/compare/v3.10.0...v3.11.0) (2024-05-15)
+
+
+### Features
+
+* add option to not fail if configured resource group does not exist ([#650](https://github.com/gitlabform/gitlabform/issues/650)) ([a362d8e](https://github.com/gitlabform/gitlabform/commit/a362d8e41f9b65d2debd86e529d5939cc023dbbe)), closes [#649](https://github.com/gitlabform/gitlabform/issues/649)
+* manage Project and Group Labels via gitlabform ([#751](https://github.com/gitlabform/gitlabform/issues/751)) ([f1be42b](https://github.com/gitlabform/gitlabform/commit/f1be42bd245237f512ff61de9dfb30057bdf1700))
+* use python lib for group settings ([#746](https://github.com/gitlabform/gitlabform/issues/746)) ([03ce3fa](https://github.com/gitlabform/gitlabform/commit/03ce3faf0b01cbda4ff5f9137030a8f2ac159bb0))
+
+
+### Documentation
+
+* add fixes info into changelog for release 3.10.0 ([96a39ee](https://github.com/gitlabform/gitlabform/commit/96a39eee118b6c7e00d6dd1d47d845213876ffb3))
+* move commit message guideline ([c0d3f68](https://github.com/gitlabform/gitlabform/commit/c0d3f68255fb13eb2a769d7df4a6888ffdcc742d))
+
 ## 3.10.0
 
 ### Features
 
 * Add support for Application Settings [#120](https://github.com/gitlabform/gitlabform/issues/120), PR [#742](https://github.com/gitlabform/gitlabform/pull/742) ([Tim Knight (DWP)](https://github.com/TimKnight-DWP))
-
-### Fixes
-
-* Improved error handling when processing project already transferred [#745](https://github.com/gitlabform/gitlabform/pull/745) ([long-wan-ep](https://github.com/long-wan-ep))
-
 
 ### Refactors